AIMAID-7-migration-table-people
Migration for table people where ID is auto increment and PK, Name is a string type with limit of 100 characters, NIF is a string type with limit of 9 characters and is unique (can't repeat), Birthdate is a date type, and timestamps for default predefinition of laravel. Now without the comments in portuguese.